requirements.txt
albucore==0.0.20
albumentations==1.4.21
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.38.0
Deprecated==1.2.15
docker==7.1.0
docker-pycreds==0.4.0
eval_type_backport==0.2.0
fastapi==0.115.5
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.14.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mlflow==2.18.0
mlflow-skinny==2.18.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.19.0
setproctitle==1.3.4
setuptools==75.5.0
simsimd==6.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
starlette==0.41.3
stringzilla==3.11.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.11
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wandb==0.18.7
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
